# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Error {
  error: string;
}

export type TradingOverviewMode = typeof TradingOverviewMode[keyof typeof TradingOverviewMode];


export const TradingOverviewMode = {
  paper: 'paper',
  live_locked: 'live_locked',
} as const;

export type TradingOverviewSession = typeof TradingOverviewSession[keyof typeof TradingOverviewSession];


export const TradingOverviewSession = {
  asian: 'asian',
  london: 'london',
  new_york: 'new_york',
  outside: 'outside',
} as const;

export type TradingOverviewTelegramStatus = typeof TradingOverviewTelegramStatus[keyof typeof TradingOverviewTelegramStatus];


export const TradingOverviewTelegramStatus = {
  not_configured: 'not_configured',
  ready: 'ready',
  error: 'error',
} as const;

export type TradingOverviewExchangeStatus = typeof TradingOverviewExchangeStatus[keyof typeof TradingOverviewExchangeStatus];


export const TradingOverviewExchangeStatus = {
  not_configured: 'not_configured',
  configured: 'configured',
  error: 'error',
} as const;

export interface TradingOverview {
  mode: TradingOverviewMode;
  symbol: string;
  session: TradingOverviewSession;
  equity: number;
  todayPnl: number;
  winRate: number;
  tradesToday: number;
  maxTradesPerDay: number;
  lastSync: string;
  telegramStatus: TradingOverviewTelegramStatus;
  exchangeStatus: TradingOverviewExchangeStatus;
}

export type TradingAnalysisTrend = typeof TradingAnalysisTrend[keyof typeof TradingAnalysisTrend];


export const TradingAnalysisTrend = {
  bullish: 'bullish',
  bearish: 'bearish',
  ranging: 'ranging',
  unclear: 'unclear',
} as const;

export type TradingAnalysisBias = typeof TradingAnalysisBias[keyof typeof TradingAnalysisBias];


export const TradingAnalysisBias = {
  long: 'long',
  short: 'short',
  neutral: 'neutral',
} as const;

export type TradingAnalysisAsianRange = {
  low: number;
  high: number;
  status: string;
};

export type TradingAnalysisSetup = {
  status: string;
  entry: number;
  stopLoss: number;
  target: number;
  riskReward: number;
};

export interface TradingAnalysis {
  symbol: string;
  timeframe: string;
  trend: TradingAnalysisTrend;
  bias: TradingAnalysisBias;
  confidence: number;
  session: string;
  asianRange: TradingAnalysisAsianRange;
  structure: string[];
  liquidity: string[];
  setup: TradingAnalysisSetup;
  invalidation: string[];
  checkedAt: string;
}

export type TradeSide = typeof TradeSide[keyof typeof TradeSide];


export const TradeSide = {
  long: 'long',
  short: 'short',
} as const;

export type TradeStatus = typeof TradeStatus[keyof typeof TradeStatus];


export const TradeStatus = {
  open: 'open',
  won: 'won',
  lost: 'lost',
  rejected: 'rejected',
} as const;

export type TradeAuditChecklistItem = {
  label: string;
  passed: boolean;
};

export type TradeAudit = {
  grade: string;
  verdict: string;
  reasons: string[];
  checklist: TradeAuditChecklistItem[];
};

export interface Trade {
  id: number;
  symbol: string;
  side: TradeSide;
  status: TradeStatus;
  entry: number;
  /** @nullable */
  exit?: number | null;
  stopLoss: number;
  target: number;
  /** @nullable */
  pnl?: number | null;
  riskReward: number;
  session: string;
  setup: string;
  confluences: string[];
  openedAt: string;
  /** @nullable */
  closedAt?: string | null;
  audit: TradeAudit;
}

export type PaperTradeInputSymbol = typeof PaperTradeInputSymbol[keyof typeof PaperTradeInputSymbol];


export const PaperTradeInputSymbol = {
  'BTC/USDT': 'BTC/USDT',
  'ETH/USDT': 'ETH/USDT',
} as const;

export type PaperTradeInputSide = typeof PaperTradeInputSide[keyof typeof PaperTradeInputSide];


export const PaperTradeInputSide = {
  long: 'long',
  short: 'short',
} as const;

export interface PaperTradeInput {
  symbol?: PaperTradeInputSymbol;
  side: PaperTradeInputSide;
  entry: number;
  stopLoss: number;
  target: number;
  session: string;
  setup: string;
}

export interface CloseTradeInput {
  exit: number;
  verdict: string;
}

export type BotControlMode = typeof BotControlMode[keyof typeof BotControlMode];


export const BotControlMode = {
  paper: 'paper',
  live_locked: 'live_locked',
} as const;

export type BotControlTelegramStatus = typeof BotControlTelegramStatus[keyof typeof BotControlTelegramStatus];


export const BotControlTelegramStatus = {
  not_configured: 'not_configured',
  ready: 'ready',
  error: 'error',
} as const;

export type BotControlExchangeStatus = typeof BotControlExchangeStatus[keyof typeof BotControlExchangeStatus];


export const BotControlExchangeStatus = {
  not_configured: 'not_configured',
  configured: 'configured',
  error: 'error',
} as const;

export interface BotControl {
  mode: BotControlMode;
  running: boolean;
  liveOrdersEnabled: boolean;
  symbol: string;
  maxTradesPerDay: number;
  riskPerTrade: number;
  sessions: string[];
  telegramStatus: BotControlTelegramStatus;
  exchangeStatus: BotControlExchangeStatus;
}

export interface BotControlUpdate {
  running?: boolean;
  /**
     * @minimum 1
     * @maximum 5
     */
  maxTradesPerDay?: number;
  /**
     * @minimum 0.1
     * @maximum 2
     */
  riskPerTrade?: number;
}

export type AlertType = typeof AlertType[keyof typeof AlertType];


export const AlertType = {
  setup: 'setup',
  trade_opened: 'trade_opened',
  trade_closed: 'trade_closed',
  audit: 'audit',
  system: 'system',
} as const;

export type AlertStatus = typeof AlertStatus[keyof typeof AlertStatus];


export const AlertStatus = {
  sent: 'sent',
  pending: 'pending',
  failed: 'failed',
  not_configured: 'not_configured',
} as const;

export interface Alert {
  id: number;
  type: AlertType;
  title: string;
  body: string;
  status: AlertStatus;
  createdAt: string;
}

export type ListTradingTradesParams = {
status?: ListTradingTradesStatus;
};

export type ListTradingTradesStatus = typeof ListTradingTradesStatus[keyof typeof ListTradingTradesStatus];


export const ListTradingTradesStatus = {
  all: 'all',
  open: 'open',
  won: 'won',
  lost: 'lost',
  rejected: 'rejected',
} as const;