# lib/api-zod/src/generated/types/alert.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { AlertStatus } from './alertStatus';
import type { AlertType } from './alertType';

export interface Alert {
  id: number;
  type: AlertType;
  title: string;
  body: string;
  status: AlertStatus;
  createdAt: string;
}