# lib/api-zod/src/generated/types/botControlUpdate.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface BotControlUpdate {
  running?: boolean;
  /**
     * @minimum 1
     * @maximum 5
     */
  maxTradesPerDay?: number;
  /**
     * @minimum 0.1
     * @maximum 2
     */
  riskPerTrade?: number;
}